requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4
keras==2.13.1
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.4.0
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytz==2022.7.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-hub==0.14.0
tensorflow-macos==2.13.0
termcolor==2.3.0
typing_extensions==4.5.0
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
